separate matrix.js into multiple other files
